[Update] Bonus correction printing death

+++ philo_bonus/reap_bonus.c
@@ -6,7 +6,7 @@
 /*   By: cado-car <cado-car@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/23 12:35:23 by cado-car          #+#    #+#             */
-/*   Updated: 2023/03/23 18:45:13 by cado-car         ###   ########.fr       */
+/*   Updated: 2023/03/23 19:36:18 by cado-car         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ void	*reap(void *table_ptr)
 	sem_wait(table->args.death);
 	i = -1;
 	while (++i < table->args.nb_philo)
-		// if (table->args.pid_array[i] != -1)
-		kill(table->args.pid_array[i], SIGKILL);
+		if (table->args.pid_array[i] != -1)
+			kill(table->args.pid_array[i], SIGKILL);
 	return (NULL);
 }
